requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0